chore: Make test/pipe.mjs
It is used to test the passing of binary files to pipes.
